refactor allow_test
